+++ packages/node-common/src/FileSystemRawTreeSitterQueryProvider.ts
